Extracting `RuleInfo` to a new class; Extracting `CollectTokenSortsAndAttributes` to a new class and make it a compiler pass; Creating `tokenSorts` and `rulePriorities` in module.
