+++ haskell-backend/src/main/java/org/kframework/backend/haskell/HaskellRewriter.java
@@ -348,7 +348,7 @@ public class HaskellRewriter implements Function<Definition, Rewriter> {
 
             @Override
             public RewriterResult prove(Module rules, Boolean reuseDef) {
-                Module kompiledModule = KoreBackend.getKompiledModule(module, true, kompileOptions);
+                Module kompiledModule = KoreBackend.getKompiledModule(module, true, kompileOptions, false);
                 ModuleToKORE converter = new ModuleToKORE(kompiledModule, def.topCellInitializer, kompileOptions, kem);
                 String defPath = reuseDef ? files.resolveKompiled("definition.kore").getAbsolutePath() : saveKoreDefinitionToTemp(converter);
                 String specPath = saveKoreSpecToTemp(converter, rules);
